Add requested changes
