+++ starknet_devnet/blueprints/rpc/storage.py
@@ -2,7 +2,7 @@
 RPC storage endpoints
 """
 
-from starknet_devnet.blueprints.rpc.rpc_utils import assert_block_id_is_latest, rpc_felt
+from starknet_devnet.blueprints.rpc.utils import assert_block_id_is_latest, rpc_felt
 from starknet_devnet.blueprints.rpc.structures.types import Address, BlockId, Felt, RpcError
 from starknet_devnet.state import state
